requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
scikit-learn==1.3.2
joblib==1.3.2
psycopg2-binary==2.9.9
pandas==2.1.3
